requirements.txt
haversine==2.8.0
plotly==5.15.0
folium==0.14.0
pandas==2.0.3
streamlit==1.25.0
Pillow==9.5.0
streamlit-folium==0.13.0
numpy==1.24.4
emoji==2.8.0